feat(GoogleDriveVideoAdder): done for today
Played around with ffmpeg commands, tried to concatenate the videos together. So far, the best solution is to use the file method (https://trac.ffmpeg.org/wiki/Concatenate). Another solution can be to basically not generate a file but to use weird syntax to add it as an argument, it simulating it to be a file.